refactor(front-end): move EditClient para Register
