+++ src/pages/client/components/styles.scss
@@ -63,7 +63,6 @@
     overflow: hidden;
   
     &::-webkit-scrollbar-button {
-      height: 0;
       display: none;
     }
   
@@ -84,7 +83,6 @@
       }
     }
   }
-  
 
   .user {
     display: flex;
